+++ AddressIdentifier.hpp
@@ -29,8 +29,11 @@ class SensorsIdentifierManager {
         delete[] numEnumSensorInVectorArray;
     }
 
-    void addSensor(uint8_t address, std::map<uint8_t, Sensor*> * sensors, std::vector<csa::ConflictingAddressStruct*> * conflict);
-    void addSensor(unsigned int enumPos, uint8_t address, std::map<uint8_t, Sensor*> * sensors, std::vector<csa::ConflictingAddressStruct*> * conflict);
+    void addSensor(uint32_t address, std::map<uint32_t , Sensor*> * sensors,
+                   std::vector<csa::ConflictingAddressStruct*> * conflict);
+    void addSensor(unsigned int enumPos, uint32_t address,
+                   std::map<uint32_t,Sensor*> * sensors,
+                   std::vector<csa::ConflictingAddressStruct*> * conflict);
 
   private:
     std::vector<unsigned int> * numEnumSensorInVectorArray;
@@ -38,7 +41,7 @@ class SensorsIdentifierManager {
     void JsonObjectToArrOfVectors(JsonDocument* doc);
 
 
-    static Sensor *getSensorPointerForEnumPos(unsigned int enumPos, uint8_t address);
+    static Sensor *getSensorPointerForEnumPos(unsigned int enumPos, uint32_t address);
 };
